Added redirection on successful login
