+++ src/services/ContameHttpClient.js
@@ -12,10 +12,7 @@ export class HttpClient {
 
   async request(url, method, data, headers) {
     // Hack!!
-    if (this.token === null) {
-      const loginInfo = await signIn();
-      this.token = loginInfo.data.token;
-    }
+
     headers = headers || {};
     headers = {...headers, Authorization: `Bearer ${this.token}`};
     return await axios.request({url, method, data, headers});
